+++ common/src/main/java/com/timmie/mightyarchitect/control/compose/planner/GroundPlanningToolBase.java
@@ -12,7 +12,6 @@ import com.timmie.mightyarchitect.foundation.utility.outliner.Outline.OutlinePar
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.player.LocalPlayer;
 import net.minecraft.core.BlockPos;
-import net.minecraft.core.Vec3i;
 import net.minecraft.util.Mth;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.item.context.BlockPlaceContext;
@@ -53,8 +52,7 @@ public abstract class GroundPlanningToolBase extends ComposerToolBase {
 		if (trace != null && trace.getType() == Type.BLOCK) {
 
 			var hitVec = trace.getLocation();
-			var hitVec3i = new Vec3i(Mth.floor(hitVec.x), Mth.floor(hitVec.y), Mth.floor(hitVec.z));
-			BlockPos hit = new BlockPos(hitVec3i);
+			BlockPos hit = new BlockPos(Mth.floor(hitVec.x), Mth.floor(hitVec.y), Mth.floor(hitVec.z));
 			makeStacksTransparent(player, hit);
 
 			//? if >=1.20 {