+++ common/src/main/java/com/timmie/mightyarchitect/gui/PalettePickerScreen.java
@@ -13,13 +13,9 @@ import com.timmie.mightyarchitect.control.palette.PaletteDefinition;
 import com.timmie.mightyarchitect.control.palette.PaletteStorage;
 import com.timmie.mightyarchitect.control.storage.ArchitectPaths;
 import com.timmie.mightyarchitect.foundation.utility.FilesHelper;
+import com.timmie.mightyarchitect.foundation.utility.Mouse;
 import com.timmie.mightyarchitect.gui.widgets.IconButton;
 import net.minecraft.ChatFormatting;
-//? if >=1.21.11 {
-import net.minecraft.util.Util;
-//?} else {
-/*import net.minecraft.Util;
-*///?}
 import net.minecraft.client.Minecraft;
 //? if >=26 {
 import net.minecraft.client.gui.GuiGraphicsExtractor;
@@ -230,9 +226,9 @@ public class PalettePickerScreen extends AbstractSimiScreen {
 
 			if (guibutton.isMouseOver(mouseX, mouseY)) {
 				guibutton.playDownSound(this.minecraft.getSoundManager());
-				if (mouseButton == 0)
+				if (mouseButton == Mouse.LEFT)
 					this.buttonClicked(guibutton);
-				if (mouseButton == 1)
+				if (mouseButton == Mouse.RIGHT)
 					this.buttonRightClicked(guibutton);
 				return true;
 			}
@@ -248,8 +244,7 @@ public class PalettePickerScreen extends AbstractSimiScreen {
 		if (button == buttonOpenFolder) {
 			Path folder = ArchitectPaths.palettes();
 			FilesHelper.createFolderIfMissing(folder);
-			Util.getPlatform()
-				.openFile(folder.toFile());
+			McCompat.openFolder(folder);
 		}
 
 		if (button == buttonRefresh) {